fix(deps): add missing optional peer dependencies
A couple of optional peer dependencies were not declared in the
package.json. This prevents compatibility with modern versions of yarn,
since yarn will block access to undecalred dependencies.

This adds the missing peer dependencies and marks them as optional so
they only need to be installed when actually required